% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvs-package.r
\docType{package}
\name{mvs-package}
\alias{mvs-package}
\title{mvs: Methods for High-Dimensional Multi-View Learning.}
\description{
Methods for high-dimensional multi-view learning based on the multi-view stacking (MVS) framework. For technical details on the MVS and StaPLR methods see <doi:10.1016/j.inffus.2020.03.007> and <doi:10.3389/fnins.2022.830630>.
}
\details{
Details
}
\author{
Wouter van Loon [cre, aut]  <\email{w.s.van.loon@fsw.leidenuniv.nl}>

Marjolein Fokkema [ctb]
}
